requirements.txt
numpy==1.24.2
Pillow==9.4.0
streamlit==1.20.0
tensorflow==2.11.0

